feat(leaflet): Remove internal leaflet and react-leaflet usage.
The motivation for this is to allow supporting leaflet 0.x and 1.x at the same time. It also became

apparent that it was trivial to create a Leaflet GridLayer to display any of the visualizations so

it removes the custom react-leaflet layers from the library also. (see example.js) This also

separates the Choropleth break functions into their own files.

BREAKING CHANGE: No longer extends leaflet layers or react-leaflet layers. Must be done manually.